+++ python/sdist/amici/jax/__init__.py
@@ -10,6 +10,7 @@ compatibility.
 from warnings import warn
 
 from amici.jax.model import JAXModel
+from amici.jax.nn import Flatten, cat, generate_equinox, tanhshrink
 from amici.jax.petab import (
     JAXProblem,
     ReturnValue,
@@ -26,7 +27,11 @@ warn(
 __all__ = [
     "JAXModel",
     "JAXProblem",
+    "Flatten",
+    "generate_equinox",
     "run_simulations",
     "petab_simulate",
     "ReturnValue",
+    "tanhshrink",
+    "cat",
 ]