+++ python/sdist/amici/petab/petab_import.py
@@ -7,9 +7,11 @@ into AMICI.
 
 import logging
 import os
+import re
 import shutil
 from pathlib import Path
 
+import pandas as pd
 import petab.v1 as petab
 from petab.v1.models import MODEL_TYPE_PYSB, MODEL_TYPE_SBML
 
@@ -88,12 +90,6 @@ def import_petab_problem(
             "Unsupported model type " + petab_problem.model.type_id
         )
 
-    if petab_problem.mapping_df is not None:
-        # It's partially supported. Remove at your own risk...
-        raise NotImplementedError(
-            "PEtab v2.0.0 mapping tables are not yet supported."
-        )
-
     model_name = model_name or petab_problem.model.model_id
 
     if petab_problem.model.type_id == MODEL_TYPE_PYSB and model_name is None:
@@ -135,6 +131,113 @@ def import_petab_problem(
             shutil.rmtree(model_output_dir)
 
         logger.info(f"Compiling model {model_name} to {model_output_dir}.")
+
+        if "sciml" in petab_problem.extensions_config:
+            from petab_sciml.standard import NNModelStandard
+
+            config = petab_problem.extensions_config["sciml"]
+            # TODO: only accept YAML format for now
+            hybridizations = [
+                pd.read_csv(hf, sep="\t")
+                for hf in config["hybridization_files"]
+            ]
+            hybridization_table = pd.concat(hybridizations)
+
+            input_mapping = dict(
+                zip(
+                    hybridization_table["targetId"],
+                    hybridization_table["targetValue"],
+                )
+            )
+            output_mapping = dict(
+                zip(
+                    hybridization_table["targetValue"],
+                    hybridization_table["targetId"],
+                )
+            )
+            observable_mapping = dict(
+                zip(
+                    petab_problem.observable_df["observableFormula"],
+                    petab_problem.observable_df.index,
+                )
+            )
+            hybridization = {
+                net_id: {
+                    "model": NNModelStandard.load_data(
+                        Path(net_config["location"])
+                    ),
+                    "input_vars": [
+                        input_mapping[petab_id]
+                        for petab_id, model_id in petab_problem.mapping_df.loc[
+                            petab_problem.mapping_df[petab.MODEL_ENTITY_ID]
+                            .str.split(".")
+                            .str[0]
+                            == net_id,
+                            petab.MODEL_ENTITY_ID,
+                        ]
+                        .to_dict()
+                        .items()
+                        if model_id.split(".")[1].startswith("input")
+                        and petab_id in input_mapping.keys()
+                    ],
+                    "output_vars": {
+                        output_mapping[petab_id]: _get_net_index(model_id)
+                        for petab_id, model_id in petab_problem.mapping_df.loc[
+                            petab_problem.mapping_df[petab.MODEL_ENTITY_ID]
+                            .str.split(".")
+                            .str[0]
+                            == net_id,
+                            petab.MODEL_ENTITY_ID,
+                        ]
+                        .to_dict()
+                        .items()
+                        if model_id.split(".")[1].startswith("output")
+                        and petab_id in output_mapping.keys()
+                    },
+                    "observable_vars": {
+                        observable_mapping[petab_id]: _get_net_index(model_id)
+                        for petab_id, model_id in petab_problem.mapping_df.loc[
+                            petab_problem.mapping_df[petab.MODEL_ENTITY_ID]
+                            .str.split(".")
+                            .str[0]
+                            == net_id,
+                            petab.MODEL_ENTITY_ID,
+                        ]
+                        .to_dict()
+                        .items()
+                        if model_id.split(".")[1].startswith("output")
+                        and petab_id in observable_mapping.keys()
+                    },
+                    "frozen_layers": dict(
+                        [
+                            _get_frozen_layers(model_id)
+                            for petab_id, model_id in petab_problem.mapping_df.loc[
+                                petab_problem.mapping_df[petab.MODEL_ENTITY_ID]
+                                .str.split(".")
+                                .str[0]
+                                == net_id,
+                                petab.MODEL_ENTITY_ID,
+                            ]
+                            .to_dict()
+                            .items()
+                            if petab_id in petab_problem.parameter_df.index
+                            and petab_problem.parameter_df.loc[
+                                petab_id, petab.ESTIMATE
+                            ]
+                            == 0
+                        ]
+                    ),
+                    **net_config,
+                }
+                for net_id, net_config in config["neural_nets"].items()
+            }
+            if not jax or petab_problem.model.type_id != MODEL_TYPE_SBML:
+                raise NotImplementedError(
+                    "petab_sciml extension is currently only supported for sbml models"
+                )
+        else:
+            hybridization = None
+
         # compile the model
         if petab_problem.model.type_id == MODEL_TYPE_PYSB:
             import_model_pysb(
@@ -150,6 +253,7 @@ def import_petab_problem(
                 model_name=model_name,
                 model_output_dir=model_output_dir,
                 non_estimated_parameters_as_constants=non_estimated_parameters_as_constants,
+                hybridization=hybridization,
                 jax=jax,
                 **kwargs,
             )
@@ -181,3 +285,17 @@ def import_petab_problem(
     )
 
     return model
+
+
+def _get_net_index(model_id: str):
+    matches = re.findall(r"\[(\d+)\]", model_id)
+    if matches:
+        return int(matches[-1])
+
+
+def _get_frozen_layers(model_id):
+    layers = re.findall(r"\[(.*?)\]", model_id)
+    array_attr = model_id.split(".")[-1]
+    layer_id = layers[0] if len(layers) else None
+    array_attr = array_attr if array_attr in ("weight", "bias") else None
+    return layer_id, array_attr