+++ python/sdist/amici/petab/sbml_import.py
@@ -235,6 +235,7 @@ def import_model_sbml(
     non_estimated_parameters_as_constants=True,
     output_parameter_defaults: dict[str, float] | None = None,
     discard_sbml_annotations: bool = False,
+    hybridization: dict = None,
     jax: bool = False,
     **kwargs,
 ) -> amici.SbmlImporter:
@@ -392,10 +393,15 @@ def import_model_sbml(
             output_dir=model_output_dir,
             observation_model=observation_model,
             verbose=verbose,
+            hybridization=hybridization,
             **kwargs,
         )
         return sbml_importer
     else:
+        if hybridization:
+            raise NotImplementedError(
+                "Hybridization is currently only supported for JAX models."
+            )
         sbml_importer.sbml2amici(
             model_name=model_name,
             output_dir=model_output_dir,