+++ python/sdist/amici/petab/util.py
@@ -28,7 +28,7 @@ def get_states_in_condition_table(
 
     species_check_fun = {
         MODEL_TYPE_SBML: lambda x: _element_is_sbml_state(
-            petab_problem.sbml_model, x
+            petab_problem.model.sbml_model, x
         ),
         MODEL_TYPE_PYSB: lambda x: _element_is_pysb_pattern(
             petab_problem.model.model, x